feat: roll driver, fix validfrom/to type (#516)

+++ playwright/src/main/java/com/microsoft/playwright/options/SecurityDetails.java
@@ -34,10 +34,10 @@ public class SecurityDetails {
   /**
    * Unix timestamp (in seconds) specifying when this cert becomes valid. Optional.
    */
-  public Integer validFrom;
+  public Double validFrom;
   /**
    * Unix timestamp (in seconds) specifying when this cert becomes invalid. Optional.
    */
-  public Integer validTo;
+  public Double validTo;
 
 }
